+++ tencentcloud/vpc/v20170312/client.go
@@ -7995,7 +7995,7 @@ func NewModifyVpcEndPointServiceAttributeResponse() (response *ModifyVpcEndPoint
 }
 
 // ModifyVpcEndPointServiceAttribute
-// 修改终端节点服务属性。
+// 本接口（ModifyVpcEndPointServiceAttribute）用于修改终端节点服务属性。
 //
 // 
 //